requirements.txt
torch
torchvision
torchaudio
PyYAML
pandas
matplotlib
seaborn
